Update unstructured-data-upload.yaml
One revision to flight_type, removing "AIRSPACE_SIMULATION" as it should never apply in this context. Postflight data will be generated only for FLIGHT_SIMUALTIONS (i.e., the helicopter as a proxy for an AAM vehicle) and LIVE_FLIGHT (i.e., an actual AAM vehicle).

A revision to the "data_source" description, previously it had a typo (been instead of being). I also removed the OTHER_DATA note as that has been removed as an enum option.  The ETL process will require a specific data_source to work correctly. A new data source type will have to be arranged with the DM team for accommodation prior to uploading.